Fix #382 #383 resolvendo problemas de versão
