+++ src/Screens/AbsencesScreen.js
@@ -1,11 +1,21 @@
 import React from 'react';
 import Absences from '../Components/Absences';
+import ScreenHeader from '../Components/ScreenHeader';
+import {View} from 'react-native';
 
 export default class AbsencesScreen extends React.Component {
 
   render() {
     return (
-      <Absences type='complete'/>
+      <View style={{flex: 1}}>
+        <ScreenHeader
+          title='Abono'
+        />
+        <Absences
+          type='complete'
+          title='Selecione o período de abono'
+        />
+      </View>
     );
   }
 }
