--- src/Styles/styles.js
@@ -1,61 +1,0 @@
-
-const styles = {
-  title: {
-    margin: 5,
-    alignSelf: 'center',
-    fontSize: 15,
-    color: '#5f4b8b'
-  },
-  container1: {
-    flex: 1,
-    flexDirection: 'column',
-    justifyContent: 'space-around'
-  },
-  date: {
-    color: '#5f4b8b',
-    fontSize: 20
-  },
-  dates: {
-    flex: 1,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    marginRight: 5,
-    marginLeft: 5,
-    marginTop: 10,
-    borderColor: '#5f4b8b',
-    borderWidth: 2,
-    borderRadius: 5
-  },
-  justify: {
-    flex: 1.5,
-    backgroundColor: '#ede8f7',
-    marginLeft: 5,
-    marginRight: 5,
-    marginTop: 10,
-    borderColor: '#5f4b8b',
-    borderWidth: 2,
-    borderRadius: 5
-  },
-  button: {
-    width: 200,
-    marginTop: 15,
-    marginRight: 5,
-    marginLeft: 10
-  },
-  view1: {
-    flex: 1,
-    flexDirection: 'column',
-    alignItems: 'center'
-  },
-  form: {
-    width: 375
-  },
-  container2: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#fff'
-  }
-};
-
-export default styles;
